+++ daskms/experimental/arrow/reads.py
@@ -1,10 +1,10 @@
 from collections import defaultdict
 import json
-import logging
 from pathlib import Path
 import re
 from threading import Lock
 import weakref
+import warnings
 
 import dask.array as da
 from dask.array.core import normalize_chunks
@@ -19,6 +19,7 @@ from daskms.experimental.arrow.extension_types import TensorType
 from daskms.experimental.arrow.require_arrow import requires_arrow
 from daskms.constants import DASKMS_PARTITION_KEY
 from daskms.patterns import Multiton
+from daskms.utils import natural_order
 
 try:
     import pyarrow as pa
@@ -52,8 +53,6 @@ else:
 
 _parquet_table_lock = Lock()
 _parquet_table_cache = weakref.WeakValueDictionary()
-
-log = logging.getLogger(__name__)
 
 
 def natural_order(key):
@@ -143,8 +142,8 @@ def partition_chunking(partition, fragment_rows, chunks):
         unhandled_dims = set(partition_chunks.keys()) - {"row"}
 
         if len(unhandled_dims) != 0:
-            log.warning("%s chunking dimensions are currently "
-                        "ignored for arrow", unhandled_dims)
+            warnings.warn(f"{unhandled_dims} chunking dimensions are "
+                          f"currently ignored for arrow", UserWarning)
 
         # Get any user specified row chunking, defaulting to
         row_chunks = partition_chunks.get("row", fragment_rows)
@@ -267,6 +266,7 @@ def xds_from_parquet(store, columns=None, chunks=None):
                                     column, None,
                                     start, None,
                                     end, None,
+                                    adjust_chunks={"row": end - start},
                                     new_axes=new_axes,
                                     meta=meta)