+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 X.Y.Z (YYYY-MM-DD)
 ------------------
+* Ensure Natural Ordering for parquet files (:pr:`183)
+* Fix xds_from_zarr and xds_from_parquet chunking behaviour (:pr:`182`)
 * Add LazyProxy and LazyProxyMultiton patterns to dask-ms (:pr:`177`)
 * Support cloud native storage formats via fsspec (:pr:`174`)